# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UsersController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient registered successfully."
          },
          "400": {
            "description": "Invalid input data."
          }
        },
        "summary": "Register a new patient",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully."
          },
          "401": {
            "description": "Invalid email or password."
          }
        },
        "summary": "Log in to the system",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/queue": {
      "post": {
        "operationId": "QueueController_createQueueToken",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQueueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Queue token created"
          }
        },
        "summary": "Create a new queue token",
        "tags": [
          "Queue"
        ]
      },
      "get": {
        "operationId": "QueueController_getAllQueueItems",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all queue items"
          }
        },
        "summary": "Get all queue items",
        "tags": [
          "Queue"
        ]
      }
    },
    "/queue/user/{userId}": {
      "get": {
        "operationId": "QueueController_getQueueByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get queue items for a user",
        "tags": [
          "Queue"
        ]
      }
    },
    "/queue/{doctorId}": {
      "get": {
        "operationId": "QueueController_getQueueByDoctorId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "Doctor ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get queue items for a doctor",
        "tags": [
          "Queue"
        ]
      }
    },
    "/queue/{id}": {
      "put": {
        "operationId": "QueueController_updateQueueStatus",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Queue ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQueueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update status of a queue token",
        "tags": [
          "Queue"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDBITS Healthcare API",
    "description": "REST API for the MEDBITS Healthcare Management System.\n\n**Authentication**: Pass the user role in the `role` request header.\n\nValid roles: `patient` | `doctor` | `frontdesk` | `admin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "dob": {
            "type": "string",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "bloodGroup": {
            "type": "string",
            "example": "O+"
          },
          "guardianName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "dob",
          "gender",
          "bloodGroup",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "patient@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateQueueDto": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "string",
            "example": "DOC001"
          },
          "userId": {
            "type": "string",
            "example": "PAT001"
          }
        },
        "required": [
          "doctorId",
          "userId"
        ]
      },
      "UpdateQueueDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "waiting",
            "enum": [
              "waiting",
              "in-consultation",
              "done"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}